requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo==4.14.1


###-e .
